dashboard_api/requirements.txt
fastapi
uvicorn[standard]
pandas
google-generativeai>=0.5.0
python-dotenv